requirements.txt
aiohttp==3.7.4
asgiref==3.5.2
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
bson==0.5.10
certifi==2019.9.11
chardet==3.0.4
click==8.0.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
Django==4.0.5
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipython==8.4.0
jedi==0.18.1
jprq==2.1.0
matplotlib-inline==0.1.3
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
rich==10.14.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
typing_extensions==4.2.0
wcwidth==0.2.5
websockets==9.1
yarl==1.7.2
